task_diag: Fix 8-byte alignment for vma and vma_stats
Prior commit undid the 8-byte alignment for vma's. Add a comment
as to why the copy is needed. Similarly, use a temp variable for the
vma_stats.

In both cases there is no assurance that the vma or vma_stat pointer
will be 8-byte aligned, so the temp variable and a copy is a necessary
evil.

Signed-off-by: David Ahern <dsahern@gmail.com>