net: dev: Allow devices to be invisible
There are a number of use cases where it is desirable to hide netdevices
from link dumps. This patch adds the IFF_INVISIBLE flag as a way of
marking such devices. The flag can be set or reset at any time,
e.g., ip link set NAME [invisible|visible]

Intent is to hide "control" devices from a user where operations on them
can cause networking to become inoperable. It is not intended as a way of
hiding network functionality. To that end, if addresses or neighbor
entries are added for the device they will show up. Further, the device
can be referenced directly (ip li sh dev NAME or ifconfig NAME)

When a device is marked as invisible, it will not appear in link dumps
(using rtnetlink or ioctl) or in /proc/net/dev file.

TO-DO:
- invisible devices can not have an address assigned
  --> hiding routes and neighbors means hiding core network functionality
  + fail assignment attempts

- make /sys invisible for device
  + no stats or controls relevant for this device
  + disables ability to toggle attribute - must be set at create time

Signed-off-by: David Ahern <dsa@cumulusnetworks.com>