net: Add netdevice operation for incrementing packet counters
Add ndo_update_stats operation for updating stats when they are
managed through a custom implementation.

Add dev_update_stats helper to invoke the new NDO when it is
implemented and to update standard net_device_stats when it is
not.

For now, allow RX, TX, RX dropped and TX dropped to be incremented
via the new handler.

This will be used by XDP programs to manage stats on upper layer
devices.

Signed-off-by: David Ahern <dsahern@gmail.com>