ipv6: Move gateway checks to per nexthop
With nexthop objects, the gateway is not per fib6_info entry but
per-fib6_nh embedded within the nexthop. Furthermore, the nexthop
spec can be updated, flipping between a gateway and device only.

Add a new fib_nh_has_gw flag to fib6_nh and convert references to
RTF_GATEWAY to the new flag. For IPv6 address the flag is cheaper
than checking that nh_gw is non-0 like IPv4 does.

Signed-off-by: David Ahern <dsahern@gmail.com>