ipv6: Refactor fib6_ignore_linkdown
fib6_ignore_linkdown take a fib6_info but only looks at the net_device
and its IPv6 config. Change it to take a net_device over a fib6_info as
its input argument.

In addition, move it to a header file to make the check inline and usable
later with IPv4 code without going through the ipv6 stub, and rename to
ip6_ignore_linkdown since it is only checking the setting based on the
ipv6 struct.

Signed-off-by: David Ahern <dsahern@gmail.com>