neighbor: Add skip_cache argument to neigh_output
A later patch allows an IPv6 gateway with an IPv4 route. The neighbor
entry will existing the v6 table and the cached header will contain
the ipv6 protocol which is wrong for an IPv6 packet. For an IPv4
route to use the v6 neighbor entry neigh_output needs to skip the
cached header and just use the output callback for the neigh entry.

Signed-off-by: David Ahern <dsahern@gmail.com>