ipv4: Add function to send route updates
When a nexthop spec is updated via replace, the routes linked to it are
implicitly changed. Add a function, fib_info_notify_update, to walk the
fib and send RTM_NEWROUTE notifications with NLM_F_REPLACE set for
entries linked to a fib_info that is linked to the nexthop that has
changed.

The notifications are needed for legacy apps that do not understand
the new nexthop object. It really goes against the whole intent of
the new design by generating a flurry of notifications. Apps that
are nexthop aware can use the RTA_NH_ID attribute in the route
notification to just ignore it. In the future this should be
wrapped in a sysctl to allow OS'es that are fully updated to
avoid the notificaton storm.

Signed-off-by: David Ahern <dsahern@gmail.com>