ipv6: export function to send route updates
Add fib6_rt_update to send RTM_NEWROUTE with NLM_F_REPLACE set. This
helper will be used by the nexthop code to notify userspace of that
routes are impacted when a nexthop spec is updated via replace.

This notification is needed for legacy apps that do not understand
the new nexthop object. It really goes against the whole intent of
the new design by generating a flurry of notifications. Apps that
are nexthop aware can use the RTA_NH_ID attribute in the route
notification to just ignore it. In the future this should be
wrapped in a sysctl to allow OS'es that are fully updated to
avoid the notificaton storm.

Signed-off-by: David Ahern <dsahern@gmail.com>