net: Initial nexthop code
Barebones start point for nexthops. Implementation for RTM commands,
notifications, management of rbtree for holding nexthops by id, and
data structures for nexthops and nexthop config.

Similar to routes, nexthops are configured per namespace, so add
netns_nexthop struct and add it to struct net. Nexthops are maintained
in an rbtree sorted by id.

Nexthop notifications are sent when a nexthop is added or deleted,
but NOT if the delete is due to a device event or network namespace
teardown (which also involves device events). Applications are
expected to use the device down event to flush nexthops and any
routes used by the nexthops.

Signed-off-by: David Ahern <dsahern@gmail.com>