ipv6: prepare for nexthop references in a fib6_info
Move the helpers accessing fib6_nh entries in a fib6_info to nexthop.h.

fib6_info_has_gw has only 1 user; drop the helper in favor of the
builtin check. The next patch adds a check on whether the fib entry
uses a nexthop object in which case the 1 use case is always for the
built in fib6_nh.

Signed-off-by: David Ahern <dsahern@gmail.com>