ipv6: Add support for nexthop object in a fib6_info
Add struct nexthop to fib6_info. Similar to IPv4, add nh_list list to
fib6_info to manage nexthop <-> fib6_info relationship via fi6_list in
nexthops. A fib6_info that references a nexthop object can not be part
of a multipath route via the siblings hook, so nh_list can be a union
with fib6_siblings list_head saving memory.

Update the fib6_info references to check for nh and take a different path
as needed.

When a nexthop is removed, walk the f6i_list and delete the route
from the FIB.

Signed-off-by: David Ahern <dsahern@gmail.com>