ipv4: Adjust rtable entries for gateway
To allow the gateway to be either an IPv4 or IPv6 address, remove
rt_uses_gateway from rtable and replace with rt_gw_family. If
rt_gw_family is set it implies rt_uses_gateway. Rename rt_gateway
to rt_gw4 to allow smaller names when IPv6 support is added.

Signed-off-by: David Ahern <dsahern@gmail.com>