ipv4: Change for_nexthops to work with fib_nh_common
Change the for_nexthops macro to work with fib_nh_common and update the
loops to access nh_common fields. The only need for fib_nh is to access
tclassid if the feature is enabled. For that, container_of can be used to
convert the common struct to an ipv4 struct.

During the conversion, update references to fib_nh_gw4 only when the
family is AF_INET.

Signed-off-by: David Ahern <dsahern@gmail.com>