ipv4: Add support for nexthop object to fib_info
Add 'struct nexthop' to fib_info. Add nh_list list as well to manage
nexthop <-> fib_info relationship via fi_list in nexthops. This allows
quick correlation to affected fib_info entries when a nexthop is
modified or deleted.

Update the fib_info functions to check for fi->nh and take a different
path as needed.

When a nexthop is removed, walk the fi_list and mark the fib_info
entries using it as dead and then call fib_flush to remove them.

Signed-off-by: David Ahern <dsahern@gmail.com>