ipv4: Allow ipv6 gateway with ipv4 routes
Add support for RTA_VIA and allow an IPv6 nexthop for v4 routes:
   $ ip ro add 172.16.1.0/24 via inet6 2001:db8::1 dev eth0
   $ ip ro ls
   ...
   172.16.1.0/24 via inet6 2001:db8::1 dev eth0

For convenience and simplicity, userspace can use RTA_VIA to specify
AF_INET or AF_INET6 gateway.

The common fib_nexthop_info dump function uses the for_ipv4 argument
to know if the nexthop data is for an ipv4 route. If so, an ipv6
gateway is encoded as RTA_VIA as opposed to RTA_GATEWAY.

Signed-off-by: David Ahern <dsahern@gmail.com>